--- WindowDebugger/Themes/Core.xaml
@@ -1,33 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-    
-    <Color x:Key="Color.Theme">#1570a6</Color>
-    <Color x:Key="Color.Theme.Light1">#3396cf</Color>
-    <Color x:Key="Color.Theme.Dark1">#065280</Color>
-    <Color x:Key="Color.Foreground">Black</Color>
-    <Color x:Key="Color.Foreground.Disable">#999999</Color>
-    <Color x:Key="Color.Background">White</Color>
-    <Color x:Key="Color.Theme.Foreground">White</Color>
-    <Color x:Key="Color.Theme.Foreground.Disable">#6FA7C8</Color>
-    <Color x:Key="Color.Theme.Background">#232323</Color>
-    <Color x:Key="Color.Background.Header">#323232</Color>
-    <Color x:Key="Color.Background.Items">#262526</Color>
-    <Color x:Key="Color.Background.Footer">#323232</Color>
-    <Color x:Key="Color.Background.Float">#3f3f42</Color>
-    <Color x:Key="Color.Background.Control">#6f6f73</Color>
-    <SolidColorBrush x:Key="Brush.Theme" Color="{StaticResource Color.Theme}" />
-    <SolidColorBrush x:Key="Brush.Theme.Light1" Color="{StaticResource Color.Theme.Light1}" />
-    <SolidColorBrush x:Key="Brush.Theme.Dark1" Color="{StaticResource Color.Theme.Dark1}" />
-    <SolidColorBrush x:Key="Brush.Foreground" Color="{StaticResource Color.Foreground}" />
-    <SolidColorBrush x:Key="Brush.Foreground.Disable" Color="{StaticResource Color.Foreground.Disable}" />
-    <SolidColorBrush x:Key="Brush.Background" Color="{StaticResource Color.Background}" />
-    <SolidColorBrush x:Key="Brush.Theme.Foreground" Color="{StaticResource Color.Theme.Foreground}" />
-    <SolidColorBrush x:Key="Brush.Theme.Foreground.Disable" Color="{StaticResource Color.Theme.Foreground.Disable}" />
-    <SolidColorBrush x:Key="Brush.Theme.Background" Color="{StaticResource Color.Theme.Background}" />
-    <SolidColorBrush x:Key="Brush.Background.Header" Color="{StaticResource Color.Background.Header}" />
-    <SolidColorBrush x:Key="Brush.Background.Items" Color="{StaticResource Color.Background.Items}" />
-    <SolidColorBrush x:Key="Brush.Background.Footer" Color="{StaticResource Color.Background.Footer}" />
-    <SolidColorBrush x:Key="Brush.Background.Float" Color="{StaticResource Color.Background.Float}" />
-    <SolidColorBrush x:Key="Brush.Background.Control" Color="{StaticResource Color.Background.Control}" />
-    
-</ResourceDictionary>
