--- WindowDebugger/Themes/FDS/Button.xaml
@@ -1,67 +1,0 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:effects="clr-namespace:Walterlv.Windows.Effects;assembly=Walterlv.Themes.FluentDesign">
-
-    <ResourceDictionary.MergedDictionaries>
-        <ResourceDictionary Source="/Themes/Core.xaml" />
-    </ResourceDictionary.MergedDictionaries>
-
-    <Style TargetType="{x:Type Button}">
-        <Setter Property="FontSize" Value="14" />
-        <Setter Property="Margin" Value="4 0" />
-        <Setter Property="Padding" Value="24 0" />
-        <Setter Property="Height" Value="32" />
-        <Setter Property="BorderThickness" Value="0" />
-        <Setter Property="BorderBrush" Value="{x:Null}" />
-        <Setter Property="Foreground" Value="{StaticResource Brush.Theme.Foreground}" />
-        <Setter Property="Background" Value="{StaticResource Brush.Theme}" />
-        <Setter Property="effects:TiltEffect2D.IsEnabled" Value="True" />
-        <Setter Property="Stylus.IsPressAndHoldEnabled" Value="False" />
-        <Setter Property="Stylus.IsFlicksEnabled" Value="False" />
-        <Setter Property="Stylus.IsTapFeedbackEnabled" Value="False" />
-        <Setter Property="Stylus.IsTouchFeedbackEnabled" Value="False" />
-        <Setter Property="FocusVisualStyle" Value="{x:Null}" />
-        <Setter Property="Template">
-            <Setter.Value>
-                <ControlTemplate TargetType="Button">
-                    <Grid Background="{TemplateBinding Background}">
-                        <Border x:Name="Bd"
-                                BorderThickness="{TemplateBinding BorderThickness}"
-                                BorderBrush="{TemplateBinding BorderBrush}" />
-                        <TextBlock x:Name="ContentTextBlock" Foreground="{TemplateBinding Foreground}"
-                                   Margin="{TemplateBinding Padding}"
-                                   FontSize="{TemplateBinding FontSize}" Text="{TemplateBinding Content}"
-                                   HorizontalAlignment="Center" VerticalAlignment="Center" />
-                    </Grid>
-                    <ControlTemplate.Triggers>
-                        <MultiTrigger>
-                            <!-- When the pointer is over the button. -->
-                            <MultiTrigger.Conditions>
-                                <Condition Property="IsMouseOver" Value="True" />
-                                <Condition Property="IsStylusOver" Value="False" />
-                            </MultiTrigger.Conditions>
-                            <Setter TargetName="Bd" Property="Background" Value="#4ff4f4f4" />
-                        </MultiTrigger>
-                        <!-- When the pointer is pressed. -->
-                        <MultiTrigger>
-                            <MultiTrigger.Conditions>
-                                <Condition Property="IsPressed" Value="True" />
-                                <Condition Property="AreAnyTouchesOver" Value="False" />
-                            </MultiTrigger.Conditions>
-                            <Setter TargetName="Bd" Property="Background" Value="#4f1f1f1f" />
-                        </MultiTrigger>
-                        <!-- When the touch device is pressed. -->
-                        <MultiTrigger>
-                            <MultiTrigger.Conditions>
-                                <Condition Property="IsPressed" Value="True" />
-                                <Condition Property="AreAnyTouchesOver" Value="True" />
-                            </MultiTrigger.Conditions>
-                            <Setter TargetName="Bd" Property="Background" Value="#4f000000" />
-                        </MultiTrigger>
-                    </ControlTemplate.Triggers>
-                </ControlTemplate>
-            </Setter.Value>
-        </Setter>
-    </Style>
-
-</ResourceDictionary>
