--- WindowDebugger/Views/OperationsTab.xaml
@@ -1,36 +1,0 @@
-﻿<TabItem x:Name="tabItem" x:Class="WindowDebugger.Views.OperationsTab"
-         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-         xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-         xmlns:local="clr-namespace:WindowDebugger.Views"
-         xmlns:viewmodels="clr-namespace:WindowDebugger.ViewModels"
-         mc:Ignorable="d"
-         d:DesignHeight="500" d:DesignWidth="500" d:DataContext="{d:DesignInstance Type=viewmodels:WindowItem}" Header="{Binding Content, RelativeSource={RelativeSource Self}, Mode=OneWay}">
-    <Border Padding="10">
-        <Grid>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="*"/>
-                <RowDefinition Height="*"/>
-            </Grid.RowDefinitions>
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="*"/>
-                <ColumnDefinition Width="*"/>
-            </Grid.ColumnDefinitions>
-            <Button Grid.Column="0" Grid.Row="0" MaxWidth="200" Click="ButtonSetForeground_Click">Set Foreground</Button>
-            <Button Grid.Column="1" Grid.Row="0" MaxWidth="200" Click="ButtonCloseWindow_Click">Close Window</Button>
-            <Button Grid.Column="0" Grid.Row="1" MaxWidth="200" Click="ButtonRedrawWindow_Click">Redraw Window</Button>
-            <Button Grid.Column="1" Grid.Row="1" MaxWidth="200" Click="ButtonFlashWindow_Click">Flash Window</Button>
-            <Button Grid.Column="0" Grid.Row="2" MaxWidth="200" Click="ButtonKillProcess_Click">Kill Process</Button>
-            <Button Grid.Column="1" Grid.Row="2" MaxWidth="200" Click="ButtonForceKillProcess_Click">Force Kill Process</Button>
-            <Button Grid.Column="0" Grid.Row="3" MaxWidth="200" Click="ButtonSuspendThread_Click">Suspend Thread</Button>
-            <Button Grid.Column="1" Grid.Row="3" MaxWidth="200" Click="ButtonResumeThread_Click">Resume Thread</Button>
-            <Button Grid.Column="0" Grid.Row="4" MaxWidth="200" Click="ButtonKillThread_Click">Kill Thread</Button>
-        </Grid>
-    </Border>
-</TabItem>
