--- WindowDebugger/Views/OthersTab.xaml
@@ -1,62 +1,0 @@
-﻿<TabItem x:Class="WindowDebugger.Views.OthersTab"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:WindowDebugger.Views"
-             xmlns:win32Enums="clr-namespace:Lsj.Util.Win32.Enums;assembly=Lsj.Util.Win32"
-             xmlns:converters="clr-namespace:WindowDebugger.Converters"
-             xmlns:utilconverters="clr-namespace:Lsj.Util.WPF.Converters;assembly=Lsj.Util.WPF"
-             xmlns:markupExtensions="clr-namespace:Lsj.Util.WPF.MarkupExtensions;assembly=Lsj.Util.WPF"
-             xmlns:viewmodels="clr-namespace:WindowDebugger.ViewModels"
-         mc:Ignorable="d"
-         d:DesignHeight="500" d:DesignWidth="500" d:DataContext="{d:DesignInstance Type=viewmodels:WindowItem}" Header="{Binding Content, RelativeSource={RelativeSource Self}, Mode=OneWay}">
-    <Border Padding="10,10,10,0">
-        <Grid>
-            <Grid.Resources>
-                <Style TargetType="TextBlock">
-                    <Style.Setters>
-                        <Setter Property="VerticalAlignment" Value="Center"/>
-                    </Style.Setters>
-                </Style>
-                <Style TargetType="TextBox">
-                    <Style.Setters>
-                        <Setter Property="VerticalAlignment" Value="Center"/>
-                        <Setter Property="Height" Value="20"/>
-                    </Style.Setters>
-                </Style>
-                <Style TargetType="Button" BasedOn="{StaticResource {x:Type Button}}">
-                    <Style.Setters>
-                        <Setter Property="VerticalAlignment" Value="Center"/>
-                        <Setter Property="Height" Value="20"/>
-                        <Setter Property="Margin" Value="10,0,0,0"/>
-                        <Setter Property="Padding" Value="8 0" />
-                    </Style.Setters>
-                </Style>
-                <Style TargetType="ComboBox">
-                    <Style.Setters>
-                        <Setter Property="VerticalAlignment" Value="Center"/>
-                        <Setter Property="VerticalContentAlignment" Value="Center"/>
-                        <Setter Property="Height" Value="22"/>
-                    </Style.Setters>
-                </Style>
-                <Style TargetType="CheckBox">
-                    <Setter Property="VerticalAlignment" Value="Center"/>
-                </Style>
-            </Grid.Resources>
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="2*" MaxWidth="150"/>
-                <ColumnDefinition Width="2*"/>
-                <ColumnDefinition Width="3*"/>
-                <ColumnDefinition Width="*" MinWidth="80"/>
-            </Grid.ColumnDefinitions>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="*" MaxHeight="30"/>
-                <RowDefinition Height="*" MaxHeight="30"/>
-            </Grid.RowDefinitions>
-            <CheckBox Grid.Row="0" Grid.Column="0" Grid.ColumnSpan="3" Content="IsTouchWindow" IsChecked="{Binding IsTouchWindow, Mode=OneWay}" IsEnabled="False"/>
-            <TextBlock Grid.Row="1" Grid.Column="0" Grid.ColumnSpan="2" Text="Virtual Desktop ID"/>
-            <TextBox Grid.Row="1" Grid.Column="1" Grid.ColumnSpan="3" Text="{Binding VirtualDesktopID, Mode=OneWay}" IsEnabled="False"/>
-        </Grid>
-    </Border>
-</TabItem>
