+++ WindowDebugger/Views/Pages/AboutPage.xaml
@@ -1,0 +1,35 @@
+﻿<Page x:Class="WindowDebugger.Views.Pages.AboutPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:local="clr-namespace:WindowDebugger.Views.Pages"
+      xmlns:win11="clr-namespace:WPFUI.Controls;assembly=WPFUI"
+      mc:Ignorable="d"
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="AboutPage"
+      TextBlock.Foreground="{DynamicResource TextFillColorPrimaryBrush}"
+      TextBlock.FontSize="14"
+      Loaded="AboutPage_Loaded">
+
+    <ScrollViewer Margin="0 48 0 0">
+        <StackPanel Margin="16">
+            <TextBlock Margin="0 0 0 8" LineHeight="24">
+                <Run Text="Window Debugger" />
+                <Run x:Name="VersionRun" d:Text="1.0.0" />
+                <LineBreak />
+                <Run Text="© 2019-2022 kkwpsv. Some rights reserved." />
+            </TextBlock>
+            <win11:Hyperlink Content="Original release" Icon="Link48"
+                             Margin="-4 0 0 0" Padding="0 4 16 4"
+                             NavigateUri="https://gitlab.sdlsj.net/lsj/windowdebugger/-/releases" />
+            <win11:Hyperlink Content="Source code and issues on GitHub" Icon="Link48"
+                             Margin="-4 0 0 0" Padding="0 4 16 4"
+                             NavigateUri="https://github.com/kkwpsv/WindowDebugger" />
+            <win11:Hyperlink Content="DO WHAT THE FUCK YOU WANT TO PUBLIC LICENSE" Icon="Link48"
+                             Margin="-4 0 0 0" Padding="0 4 16 4"
+                             NavigateUri="https://github.com/kkwpsv/WindowDebugger/blob/master/LICENSE" Cursor="Hand" />
+        </StackPanel>
+    </ScrollViewer>
+
+</Page>
