+++ WindowDebugger/Views/Pages/WindowsPage.xaml
@@ -1,0 +1,112 @@
+﻿<Page x:Class="WindowDebugger.Views.Pages.WindowsPage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:local="clr-namespace:WindowDebugger.Views.Pages"
+      xmlns:converters="clr-namespace:WindowDebugger.Converters"
+      xmlns:viewmodels="clr-namespace:WindowDebugger.ViewModels"
+      xmlns:tabs="clr-namespace:WindowDebugger.Views.Tabs"
+      xmlns:win11="clr-namespace:WPFUI.Controls;assembly=WPFUI"
+      mc:Ignorable="d"
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="WindowsPage"
+      DataContext="{x:Static viewmodels:ViewModel.Instance}"
+      Loaded="Page_Loaded">
+
+    <FrameworkElement.Resources>
+        <converters:IntPtrToInt32Converter x:Key="IntPtrToInt32Converter" />
+        <DataTemplate x:Key="DataTemplate.WindowItem" DataType="viewmodels:WindowItem">
+            <Grid>
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="72" />
+                    <ColumnDefinition />
+                </Grid.ColumnDefinitions>
+                <TextBlock Grid.Column="0" Opacity="0.5" Margin="10 4"
+                           FontFamily="Consolas" FontSize="12"
+                           Text="{Binding WindowHandle, Converter={StaticResource IntPtrToInt32Converter}, StringFormat={}{0:X8}, Mode=OneWay}" />
+                <TextBlock Grid.Column="1" Margin="0 4" Text="{Binding Text, Mode=OneWay}"
+                           TextWrapping="WrapWithOverflow" TextTrimming="CharacterEllipsis" />
+            </Grid>
+        </DataTemplate>
+    </FrameworkElement.Resources>
+
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*" MinWidth="250"/>
+            <ColumnDefinition Width="4*"/>
+        </Grid.ColumnDefinitions>
+
+        <!--窗口列表-->
+        <Grid Grid.Column="0">
+            <Grid.RowDefinitions>
+                <RowDefinition Height="48"/>
+                <RowDefinition Height="Auto"/>
+                <RowDefinition Height="*"/>
+            </Grid.RowDefinitions>
+
+            <Expander Grid.Row="1" Grid.ColumnSpan="3" Margin="8 8 8 0">
+                <Expander.Header>
+                    <Grid Margin="-4 -4 8 -4">
+                        <Grid.ColumnDefinitions>
+                            <ColumnDefinition />
+                            <ColumnDefinition Width="Auto" />
+                        </Grid.ColumnDefinitions>
+                        <TextBox Grid.Row="1" Grid.Column="0"
+                                 Padding="8 4" FontSize="14" Height="32"
+                                 Text="{Binding SearchText, UpdateSourceTrigger=PropertyChanged}"
+                                 VerticalContentAlignment="Center"
+                                 ToolTip="Type the handle or the title of a window to filter the list.">
+                        </TextBox>
+                        <win11:Button Grid.Row="1" Grid.Column="1" Appearance="Secondary"
+                                      Margin="4 0 0 0" Padding="0" Width="32" Height="32"
+                                      Content="&#xE895;" FontFamily="Segoe MDL2 Assets"
+                                      ToolTip="Refresh the window list."
+                                      Click="RefreshWindowList" />
+                    </Grid>
+                </Expander.Header>
+                <StackPanel>
+                    <StackPanel.Resources>
+                        <Style TargetType="{x:Type CheckBox}" BasedOn="{StaticResource {x:Type CheckBox}}">
+                            <Setter Property="Margin" Value="0 8 0 0" />
+                        </Style>
+                    </StackPanel.Resources>
+                    <TextBlock Text="Filter Options:" />
+                    <CheckBox Content="Visible windows only" />
+                    <CheckBox Content="Titled windows only" />
+                </StackPanel>
+            </Expander>
+
+            <ListBox Grid.Row="2" Grid.ColumnSpan="3" x:Name="WindowList"
+                     ItemsSource="{Binding Windows, Mode=OneWay}"
+                     SelectedItem="{Binding SelectedWindow, Mode=TwoWay}"
+                     BorderThickness="0" Background="Transparent" Padding="0"
+                     ItemTemplate="{StaticResource DataTemplate.WindowItem}"
+                     ScrollViewer.HorizontalScrollBarVisibility="Disabled"
+                     VirtualizingPanel.ScrollUnit="Pixel" />
+        </Grid>
+
+        <Grid Grid.Row="1" Grid.Column="1">
+            <Grid.RowDefinitions>
+                <RowDefinition />
+                <RowDefinition Height="Auto" />
+            </Grid.RowDefinitions>
+            <TabControl Margin="0 20 12 8"
+                        DataContext="{Binding SelectedWindow}"
+                        DataContextChanged="TabControl_DataContextChanged"
+                        ItemContainerStyle="{StaticResource {x:Type TabItem}}">
+                <tabs:WindowInfoTab Header="Info"/>
+                <tabs:WindowStylesTab Header="Styles"/>
+                <tabs:WindowStylesExTab Header="StyleExes"/>
+                <tabs:ClassStylesTab Header="ClassStyles"/>
+                <tabs:OperationsTab Header="Operations"/>
+                <tabs:DWMTab Header="DWM"/>
+                <tabs:OthersTab Header="Others"/>
+            </TabControl>
+            <Button Grid.Row="1" Content="Reload current window info"
+                    HorizontalAlignment="Center" Click="RefreshCurrent"
+                    Margin="16,0,10,16" Padding="24 0" Height="40">
+            </Button>
+        </Grid>
+    </Grid>
+</Page>
