+++ WindowDebugger/Views/Tabs/OperationsTab.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using System.Windows.Controls;
 using WindowDebugger.ViewModels;
 
-namespace WindowDebugger.Views
+namespace WindowDebugger.Views.Tabs
 {
     /// <summary>
     /// OperationsTab.xaml 的交互逻辑
