+++ WindowDebugger/Views/Tabs/OthersTab.xaml
@@ -1,0 +1,61 @@
+﻿<TabItem x:Class="WindowDebugger.Views.Tabs.OthersTab"
+         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+         xmlns:local="clr-namespace:WindowDebugger.Views.Tabs"
+         xmlns:win32Enums="clr-namespace:Lsj.Util.Win32.Enums;assembly=Lsj.Util.Win32"
+         xmlns:converters="clr-namespace:WindowDebugger.Converters"
+         xmlns:utilconverters="clr-namespace:Lsj.Util.WPF.Converters;assembly=Lsj.Util.WPF"
+         xmlns:markupExtensions="clr-namespace:Lsj.Util.WPF.MarkupExtensions;assembly=Lsj.Util.WPF"
+         xmlns:win11="clr-namespace:WPFUI.Controls;assembly=WPFUI"
+         xmlns:viewmodels="clr-namespace:WindowDebugger.ViewModels"
+         mc:Ignorable="d"
+         d:DesignHeight="500" d:DesignWidth="500" d:DataContext="{d:DesignInstance Type=viewmodels:WindowItem}" Header="{Binding Content, RelativeSource={RelativeSource Self}, Mode=OneWay}">
+    <ScrollViewer Padding="12 12 12 0">
+        <Grid>
+            <Grid.Resources>
+                <Style TargetType="TextBlock">
+                    <Style.Setters>
+                        <Setter Property="VerticalAlignment" Value="Center"/>
+                    </Style.Setters>
+                </Style>
+                <Style TargetType="Button" BasedOn="{StaticResource {x:Type Button}}">
+                    <Setter Property="Padding" Value="8 0" />
+                    <Setter Property="Height" Value="32" />
+                </Style>
+                <Style TargetType="ComboBox" BasedOn="{StaticResource {x:Type ComboBox}}">
+                    <Setter Property="Margin" Value="0 4" />
+                </Style>
+                <Style TargetType="CheckBox" BasedOn="{StaticResource {x:Type CheckBox}}">
+                    <Setter Property="Margin" Value="0 4" />
+                    <Setter Property="VerticalAlignment" Value="Center"/>
+                </Style>
+                <Style TargetType="win11:ToggleSwitch" BasedOn="{StaticResource {x:Type win11:ToggleSwitch}}">
+                    <Setter Property="Margin" Value="0 4" />
+                </Style>
+            </Grid.Resources>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="2*" MaxWidth="150" SharedSizeGroup="OtherOperationHeader"/>
+                <ColumnDefinition Width="5*"/>
+                <ColumnDefinition Width="*" MinWidth="80"/>
+            </Grid.ColumnDefinitions>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
+                <RowDefinition Height="Auto" />
+            </Grid.RowDefinitions>
+            <TextBlock Grid.Row="0" Grid.Column="0" Grid.ColumnSpan="2" Text="IsTouchWindow"/>
+            <CheckBox Grid.Row="0" Grid.Column="1" IsChecked="{Binding IsTouchWindow, Mode=OneWay}" Content="{Binding IsTouchWindow, Mode=OneWay}" IsEnabled="False" />
+            <TextBlock Grid.Row="1" Grid.Column="0" Grid.ColumnSpan="2" Text="Virtual Desktop ID"/>
+            <TextBox Grid.Row="1" Grid.Column="1" Grid.ColumnSpan="3" Text="{Binding VirtualDesktopID, Mode=OneWay}" IsEnabled="False"/>
+        </Grid>
+    </ScrollViewer>
+</TabItem>
