+++ WindowDebugger/Views/Tabs/OthersTab.xaml.cs
@@ -15,7 +15,7 @@ using System.Windows.Navigation;
 using System.Windows.Shapes;
 using WindowDebugger.ViewModels;
 
-namespace WindowDebugger.Views
+namespace WindowDebugger.Views.Tabs
 {
     /// <summary>
     /// OthersTab.xaml 的交互逻辑
