+++ WindowDebugger/WindowDebugger.csproj
@@ -2,8 +2,9 @@
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFrameworks>netcoreapp3.1;net48;net5-windows;net45</TargetFrameworks>
+    <TargetFrameworks>net6.0-windows;net5.0-windows;netcoreapp3.1;net48</TargetFrameworks>
     <UseWPF>true</UseWPF>
+    <LangVersion>latest</LangVersion>
     <Version>0.0.8</Version>
     <ApplicationManifest>app.manifest</ApplicationManifest>
     <ApplicationIcon>icon.ico</ApplicationIcon>
@@ -16,7 +17,7 @@
     <PackageReference Include="Lsj.Util.WPF" Version="6.0.0" />
     <PackageReference Include="Lsj.Util.Win32.NativeUI" Version="6.0.1" />
     <PackageReference Include="System.Net.Http" Version="4.3.4" />
-    <PackageReference Include="Walterlv.Themes.FluentDesign" Version="7.1.0" />
+    <PackageReference Include="WPF-UI" Version="1.2.2-prerelease136" />
   </ItemGroup>
 
   <ItemGroup>
